requirement.txt
alldata==0.2.0
arrow==1.2.0
binaryornot==0.4.4
blis==0.7.4
camelot-py==0.10.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cookiecutter==1.7.3
cryptography==35.0.0
cymem==2.0.5
distro==1.6.0
et-xmlfile==1.1.0
ghostscript==0.7
idna==3.3
Jinja2==3.0.2
jinja2-time==0.2.0
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pathy==0.6.0
# Editable Git install (pdfextractor==0.0.1) with either a deleted local remote or invalid URI:
# 'git@https://github.com/shehrozkapoor/PDFEXTRACTOR.git:shehrozkapoor/PDFEXTRACTOR.git'
-e /Users/shehrozkapoor/Desktop/pdf_converter/PDFEXTRACTOR/src
pdfminer==20191125
pdfminer.six==20211012
pdftopng==0.2.3
Pillow==8.4.0
poyo==0.5.0
preshed==3.0.5
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
PyMuPDF==1.18.19
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
tabula-py==2.3.0
tabulate==0.8.9
text-unidecode==1.3
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
